services/mint_worker/requirements.txt
Flask==2.3.2
web3==6.9.0
google-cloud-firestore==3.12.0
google-cloud-secret-manager==2.16.0
